Volcano plot for GI scores

This plot is meant to be functionally equivalent to Fig 5b (for HeLa, equivalent of Fig 3d for PC9). Volcano plot of target-level GI scores Blue points are synthetic lethal paralog GIs with GI < 0.5 and FDR < 0.1; red points are buffering paralog GIs with GI > 0.25 and FDR < 0.1.

Usage

plot_volcano(gimap_dataset, facet_rep = TRUE, reps_to_drop = "")

Arguments

gimap_dataset

A special dataset structure that is originally setup using `setup_data()` and has had gi scores calculated with `calc_gi()`.

facet_rep

Should the replicates be wrapped with facet_wrap()?

reps_to_drop

Names of replicates that should be not plotted (Optional)

Examples

if (FALSE) { # \dontrun{

gimap_dataset <- get_example_data("gimap") %>%
  gimap_filter() %>%
  gimap_annotate(cell_line = "HELA") %>%
  gimap_normalize(
    timepoints = "day"
  ) %>%
  calc_gi()

# To plot results
plot_exp_v_obs_scatter(gimap_dataset, reps_to_drop = "Day05_RepA_early")
plot_rank_scatter(gimap_dataset, reps_to_drop = "Day05_RepA_early")
plot_volcano(gimap_dataset, reps_to_drop = "Day05_RepA_early")
} # }